requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
flask-cors==4.0.0
python-dotenv==1.0.1

# COMPATIBLE Supabase SDK dependencies (no conflicts)
supabase==2.3.7
postgrest==0.15.0
httpx==0.25.0
storage3==0.7.5
gotrue==2.8.0
realtime==1.0.5

gunicorn==21.2.0
